Revert "Updated screenshots and .gitignore of HelloWorld (#291)"
This reverts commit 6de0f442fbe44ed26ace8aadbe4e242178e22186.
