+++ App/Kconfig
@@ -37,10 +37,6 @@ menu "Tactility App"
             bool "Elecrow CrowPanel Basic 5.0"
         config TT_BOARD_LILYGO_TDECK
             bool "LilyGo T-Deck"
-        config TT_BOARD_LILYGO_TDECK_PLUS
-            bool "LilyGo T-Deck Plus"
-        config TT_BOARD_LILYGO_TDECK_PRO
-            bool "LilyGo T-Deck Pro"
         config TT_BOARD_M5STACK_CORE2
             bool "M5Stack Core2"
         config TT_BOARD_M5STACK_CORES3
