+++ App/Source/Boards.h
@@ -5,7 +5,7 @@
 #include <sdkconfig.h>
 
 // Supported hardware:
-#if defined(CONFIG_TT_BOARD_LILYGO_TDECK) || defined(CONFIG_TT_BOARD_LILYGO_TDECK_PLUS) || defined(CONFIG_TT_BOARD_LILYGO_TDECK_PRO)
+#if defined(CONFIG_TT_BOARD_LILYGO_TDECK)
 #include "LilygoTdeck.h"
 #define TT_BOARD_HARDWARE &lilygo_tdeck
 #elif defined(CONFIG_TT_BOARD_CYD_2432S024C)
@@ -14,22 +14,22 @@
 #elif defined(CONFIG_TT_BOARD_CYD_2432S032C)
 #include "CYD2432S032C.h"
 #define TT_BOARD_HARDWARE &cyd_2432S032c_config
-#elif defined(CONFIG_TT_BOARD_ELECROW_CROWPANEL_ADVANCE_28)
+#elif (defined(CONFIG_TT_BOARD_ELECROW_CROWPANEL_ADVANCE_28))
 #define TT_BOARD_HARDWARE &crowpanel_advance_28
 #include "CrowPanelAdvance28.h"
-#elif defined(CONFIG_TT_BOARD_ELECROW_CROWPANEL_ADVANCE_35)
+#elif (defined(CONFIG_TT_BOARD_ELECROW_CROWPANEL_ADVANCE_35))
 #define TT_BOARD_HARDWARE &crowpanel_advance_35
 #include "CrowPanelAdvance35.h"
-#elif defined(CONFIG_TT_BOARD_ELECROW_CROWPANEL_ADVANCE_50)
+#elif (defined(CONFIG_TT_BOARD_ELECROW_CROWPANEL_ADVANCE_50))
 #define TT_BOARD_HARDWARE &crowpanel_advance_50
 #include "CrowPanelAdvance50.h"
-#elif defined(CONFIG_TT_BOARD_ELECROW_CROWPANEL_BASIC_28)
+#elif (defined(CONFIG_TT_BOARD_ELECROW_CROWPANEL_BASIC_28))
 #define TT_BOARD_HARDWARE &crowpanel_basic_28
 #include "CrowPanelBasic28.h"
-#elif defined(CONFIG_TT_BOARD_ELECROW_CROWPANEL_BASIC_35)
+#elif (defined(CONFIG_TT_BOARD_ELECROW_CROWPANEL_BASIC_35))
 #define TT_BOARD_HARDWARE &crowpanel_basic_35
 #include "CrowPanelBasic35.h"
-#elif defined(CONFIG_TT_BOARD_ELECROW_CROWPANEL_BASIC_50)
+#elif (defined(CONFIG_TT_BOARD_ELECROW_CROWPANEL_BASIC_50))
 #define TT_BOARD_HARDWARE &crowpanel_basic_50
 #include "CrowPanelBasic50.h"
 #elif defined(CONFIG_TT_BOARD_M5STACK_CORE2)
