--- Boards/LilygoTdeckPro/CMakeLists.txt
@@ -1,7 +1,0 @@
-file(GLOB_RECURSE SOURCE_FILES Source/*.c*)
-
-idf_component_register(
-    SRCS ${SOURCE_FILES}
-    INCLUDE_DIRS "Source"
-    REQUIRES Tactility esp_lvgl_port esp_lcd ST7789 GT911 PwmBacklight driver esp_adc
-)
