--- Boards/LilygoTdeckPro/Source/LilygoTdeck.h
@@ -1,5 +1,0 @@
-#pragma once
-
-#include <Tactility/hal/Configuration.h>
-
-extern const tt::hal::Configuration lilygo_tdeck;
