--- Boards/LilygoTdeckPro/Source/hal/TdeckDisplayConstants.h
@@ -1,8 +1,0 @@
-#pragma once
-
-#define TDECK_LCD_SPI_HOST SPI2_HOST
-#define TDECK_LCD_PIN_CS GPIO_NUM_12
-#define TDECK_LCD_PIN_DC GPIO_NUM_11 // RS
-#define TDECK_LCD_HORIZONTAL_RESOLUTION 320
-#define TDECK_LCD_VERTICAL_RESOLUTION 240
-#define TDECK_LCD_SPI_TRANSFER_HEIGHT (TDECK_LCD_VERTICAL_RESOLUTION / 10)
