--- Boards/LilygoTdeckPro/Source/hal/TdeckSdCard.h
@@ -1,7 +1,0 @@
-#pragma once
-
-#include "Tactility/hal/sdcard/SdCardDevice.h"
-
-using tt::hal::sdcard::SdCardDevice;
-
-std::shared_ptr<SdCardDevice> createTdeckSdCard();
