+++ Buildscripts/board.cmake
@@ -47,10 +47,6 @@ function(INIT_TACTILITY_GLOBALS SDKCONFIG_FILE)
         set(TACTILITY_BOARD_PROJECT ElecrowCrowpanelBasic50)
     elseif (board_id STREQUAL "lilygo-tdeck")
         set(TACTILITY_BOARD_PROJECT LilygoTdeck)
-    elseif (board_id STREQUAL "lilygo-tdeck-plus")
-        set(TACTILITY_BOARD_PROJECT LilygoTdeck)
-    elseif (board_id STREQUAL "lilygo-tdeck-pro")
-        set(TACTILITY_BOARD_PROJECT LilygoTdeckPro)
     elseif (board_id STREQUAL "m5stack-core2")
         set(TACTILITY_BOARD_PROJECT M5stackCore2)
     elseif (board_id STREQUAL "m5stack-cores3")
