+++ ExternalApps/HelloWorld/.gitignore
