--- sdkconfig.board.lilygo-tdeck-plus
@@ -1,53 +1,0 @@
-# Software defaults
-# Increase stack size for WiFi (fixes crash after scan)
-CONFIG_ESP_SYSTEM_EVENT_TASK_STACK_SIZE=3072
-CONFIG_LV_FONT_MONTSERRAT_14=y
-CONFIG_LV_FONT_MONTSERRAT_18=y
-CONFIG_LV_USE_USER_DATA=y
-CONFIG_LV_USE_FS_STDIO=y
-CONFIG_LV_FS_STDIO_LETTER=65
-CONFIG_LV_FS_STDIO_PATH=""
-CONFIG_LV_FS_STDIO_CACHE_SIZE=4096
-CONFIG_LV_USE_LODEPNG=y
-CONFIG_LV_USE_BUILTIN_MALLOC=n
-CONFIG_LV_USE_CLIB_MALLOC=y
-CONFIG_LV_USE_MSGBOX=n
-CONFIG_LV_USE_SPINNER=n
-CONFIG_LV_USE_WIN=n
-CONFIG_LV_USE_SNAPSHOT=y
-CONFIG_FREERTOS_HZ=1000
-CONFIG_FREERTOS_TASK_NOTIFICATION_ARRAY_ENTRIES=2
-CONFIG_FREERTOS_SMP=n
-CONFIG_FREERTOS_UNICORE=n
-CONFIG_FREERTOS_TIMER_TASK_STACK_DEPTH=4096
-CONFIG_FREERTOS_USE_TRACE_FACILITY=y
-CONFIG_PARTITION_TABLE_CUSTOM=y
-CONFIG_PARTITION_TABLE_CUSTOM_FILENAME="partitions.csv"
-CONFIG_PARTITION_TABLE_FILENAME="partitions.csv"
-CONFIG_FATFS_LFN_HEAP=y
-CONFIG_FATFS_VOLUME_COUNT=3
-
-# Hardware: Main
-CONFIG_TT_BOARD_LILYGO_TDECK_PLUS=y
-CONFIG_TT_BOARD_NAME="LilyGo T-Deck Plus"
-CONFIG_TT_BOARD_ID="lilygo-tdeck-plus"
-CONFIG_IDF_EXPERIMENTAL_FEATURES=y
-CONFIG_IDF_TARGET="esp32s3"
-CONFIG_ESP_DEFAULT_CPU_FREQ_MHZ_240=y
-CONFIG_ESP32_DEFAULT_CPU_FREQ_240=y
-CONFIG_ESPTOOLPY_FLASHSIZE_16MB=y
-CONFIG_FLASHMODE_QIO=y
-# Hardware: SPI RAM
-CONFIG_ESP32S3_SPIRAM_SUPPORT=y
-CONFIG_SPIRAM_MODE_OCT=y
-CONFIG_SPIRAM_SPEED_120M=y
-CONFIG_SPIRAM_USE_MALLOC=y
-CONFIG_SPIRAM_TRY_ALLOCATE_WIFI_LWIP=y
-# SPI Flash (can set back to 80MHz after ESP-IDF bug is resolved)
-CONFIG_ESPTOOLPY_FLASHFREQ_120M=y
-# LVGL
-CONFIG_LV_DPI_DEF=139
-CONFIG_LV_DISP_DEF_REFR_PERIOD=10
-# USB
-CONFIG_TINYUSB_MSC_ENABLED=y
-CONFIG_TINYUSB_MSC_MOUNT_PATH="/sdcard"
